Fixing CoinsCount for now
It should be +1 instead of +0.5

+++ Wide Whip/Motor2D/j1Player.h
@@ -151,7 +151,7 @@ private:
 	float				min_speed						= 0.0f;
 	float				dt_multiplied					= 0.0f;
 
-	int					coins_count						= 0;
+	float				coins_count						= 0;
 };
